core: tcp - increased the size for flags field in connection struct
(cherry picked from commit 65ed9b065c497266a4ecd9760e7c168c69b4c6e8)

+++ src/core/tcp_conn.h
@@ -199,7 +199,7 @@ struct tcp_connection{
 	struct tcp_req req; /* request data */
 	atomic_t refcnt;
 	enum sip_protos type; /* PROTO_TCP or a protocol over it, e.g. TLS */
-	unsigned short flags; /* connection related flags */
+	unsigned int flags; /* connection related flags */
 	snd_flags_t send_flags; /* special send flags */
 	enum tcp_conn_states state; /* connection state */
 	void* extra_data; /* extra data associated to the connection, 0 for tcp*/
